feat(sim): extract pure player-AI policy into tools/sim/lib/combat-policy
Estrae dist + selectPlayerAction (puri, zero stato) da tests/smoke/ai-driven-sim.js
in un modulo condiviso, re-importati senza cambio comportamento. Abilita il riuso
della stessa policy-giocatore nel full-loop combatAdapter (Option B rifinita,
consenso master-dd) senza l'accoppiamento co-op. + piano fase-0. 4/4 unit test.

Coding-Agent: claude-opus-4.8
Trace-Id: 019e88ef-754a-735c-8f36-1f8ca822c7df
