+++ tests/smoke/ai-driven-sim.js
@@ -45,6 +45,7 @@ const WebSocket = require('ws');
 const fs = require('node:fs');
 const path = require('node:path');
 const yaml = require('js-yaml');
+const { dist, selectPlayerAction } = require('../../tools/sim/combat-policy');
 
 const TUNNEL = process.env.TUNNEL;
 if (!TUNNEL) {
@@ -136,37 +137,8 @@ function attachWs(label, code, playerId, token) {
   });
 }
 
-// Manhattan distance helper (mirror of policy.js manhattanDistance).
-function dist(a, b) {
-  return Math.abs(a.x - b.x) + Math.abs(a.y - b.y);
-}
-
-// Player-side AI policy (minimal closest-enemy attack). Picks the alive
-// sistema target with smallest Manhattan distance; if in range attacks,
-// otherwise emits a single-tile move toward the target. Never spends
-// cap_pt to keep fairness budget intact.
-function selectPlayerAction(actor, units) {
-  const enemies = units.filter((u) => u.controlled_by === 'sistema' && (u.hp ?? 0) > 0);
-  if (enemies.length === 0) return null;
-  const target = enemies.sort(
-    (a, b) => dist(actor.position, a.position) - dist(actor.position, b.position),
-  )[0];
-  const range = actor.attack_range || 1;
-  if (dist(actor.position, target.position) <= range && (actor.ap_remaining ?? 0) >= 1) {
-    return { action_type: 'attack', target_id: target.id };
-  }
-  // Step one tile closer (clamped to grid bounds).
-  const dx = Math.sign(target.position.x - actor.position.x);
-  const dy = Math.sign(target.position.y - actor.position.y);
-  // Prefer larger axis to match Manhattan reduction.
-  const stepX =
-    Math.abs(target.position.x - actor.position.x) >=
-    Math.abs(target.position.y - actor.position.y);
-  const target_position = stepX
-    ? { x: actor.position.x + dx, y: actor.position.y }
-    : { x: actor.position.x, y: actor.position.y + dy };
-  return { action_type: 'move', target_position };
-}
+// `dist` + `selectPlayerAction` now live in tools/sim/lib/combat-policy.js
+// (shared with the full-loop combatAdapter). Imported at the top of this file.
 
 // 2026-05-10 — YAML scenario loader (opt-in via AI_SIM_LOAD_YAML=1).
 // Port da tools/py/batch_calibrate_non_elim.py:encounter_to_units.
